% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Replace missing values from the column \code{Ozone} of the dataset \code{airquality}}
\usage{
clean(raw)
}
\arguments{
\item{raw}{A data frame.}
}
\value{
A data frame.
}
\description{
Replace missing values from the column \code{Ozone} of the dataset \code{airquality}
}
\examples{
clean(airquality)
}
