% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lm.R
\name{fit_lm}
\alias{fit_lm}
\title{Explore changes in ozone through time with a linear model}
\usage{
fit_lm(data)
}
\arguments{
\item{data}{The \code{airquality} data frame with columns \code{Ozone} and \code{Day}.}
}
\value{
A liner model.
}
\description{
Explore changes in ozone through time with a linear model
}
\examples{
fit_lm(airquality)
}
