% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Explore changes in ozone through time with a linear model}
\usage{
linear_model(data)
}
\arguments{
\item{data}{The \code{airquality} data frame with columns \code{Ozone} and \code{Day}.}
}
\value{
A liner model.
}
\description{
Explore changes in ozone through time with a linear model
}
\examples{
linear_model(airquality)
}
