% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lm.R
\name{plot_lm}
\alias{plot_lm}
\title{Plot a linear model: \code{Ozone} ~ \code{Day}}
\usage{
plot_lm(data)
}
\arguments{
\item{data}{The \code{airquality} dataset.}
}
\value{
An object of class "ggplot".
}
\description{
Plot a linear model: \code{Ozone} ~ \code{Day}
}
\examples{
plot_lm(airquality)
}
