% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{raw_data}
\alias{raw_data}
\title{Raw \code{airquality} data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 153 rows and 6 columns.
}
\usage{
raw_data
}
\description{
Raw \code{airquality} data
}
\examples{
raw_data
}
\keyword{datasets}
