% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Read a .csv file into a tibble, quietly}
\usage{
read(path)
}
\arguments{
\item{path}{Path to a .csv file.}
}
\value{
A tibble.
}
\description{
Read a .csv file into a tibble, quietly
}
\examples{
# Setup
tmp <- tempfile()
readr::write_csv(airquality, tmp)

# Verbose
readr::read_csv(tmp)

# Quiet
read(tmp)
}
