% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a ggplot and output a path suitable for an extarnal targets file}
\usage{
save_plot(path, plot)
}
\arguments{
\item{path}{A path where to save the plot.}

\item{plot}{An object of class "ggplot".}
}
\value{
A character string.
}
\description{
Save a ggplot and output a path suitable for an extarnal targets file
}
\examples{
library(ggplot2)

plot <- ggplot(airquality, aes(Ozone, Day)) +
  geom_smooth(method = "lm")
path <- tempfile(fileext = ".png")
out <- save_plot(path, plot)
out

file.exists(out)
}
